testing/web-platform/tests/html/semantics/tabular-data/processing-model-1/span-limits-2.html

<!DOCTYPE html>
<title>Limits on rowSpan</title>
<meta name="timeout" content="long">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>

<table border=1>
  <!-- The first column must go one cell below the bottom -->
  <tr><td rowspan=65535 id=d1>a<td>
  <!-- We'll add another 65534 rows later -->
</table>

<script>
var $ = document.querySelector.bind(document);

test(() => {
    var s = "";
    for (var i = 0; i < 65532; i++) {
      s += "<tr><td>";
    }
    s += "<tr><td id=d2><tr><td>a<td>";
    document.querySelector("table").firstElementChild.innerHTML += s;
    assert_equals($("#d1").getBoundingClientRect().bottom,
                  $("#d2").getBoundingClientRect().bottom);
}, "rowspan of 65535 must be treated as 65534");
</script>
